+++ project/src/main/java/com/onlybuns/OnlyBuns/repository/Repository_Post.java
@@ -5,11 +5,11 @@ import com.onlybuns.OnlyBuns.model.Post;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
-public interface Repository_Post extends JpaRepository<Post, Integer> {
+public interface Repository_Post extends JpaRepository<Post, Long> {
     List<Post> findAll();
     List<Post> findAll(Sort sort);
     List<Post> findByParentPostIsNull();
     List<Post> findByParentPostIsNull(Sort sort);
     List<Post> findAllByAccount(Account account);
-    void deleteById(Integer id);
+    void deleteById(Long id);
 }
